Split map into layers
Map regions/background is now split into different layers which are generated at
runtime using canvas clipping/compositing. This works better for performance
reasons and for interaction/animation management. Note that this blocks
the UI thread and will perform much better if we exported these and
preloaded them as assets.