Rewrite all transaction to use transactions service
All transactions should be handled in the redux async thunks.
All functions setting up interactions with contracts should have
similiar structure and it should depend on their goal.
If they are sending transactions that have to be signed by user
then they are a bit different than requests to a blockchain reading the data.