Split out fork provider from wallet signer
This allows the fork to be used for data lookup even if the connected
wallet doesn't support the new chain; this is mostly necessary for
Tenderly forks with the Taho wallet, as Taho cannot be told to add a
fork RPC because at the moment it enforces exactly one RPC per chain id.

Note that this is still expected to fail using Taho when we get to
transaction signing, as Taho will attempt to broadcast a transaction
meant for the Tenderly fork onto Arbitrum mainnet.