Add teaser mode for public release
There will be a period during which we want to tease the dApp but not
let anyone in. For this period, the island mode `teaser` is added. This
mode hides the Feedback link, disables Connect wallet, and replaces the
center modal with one that says the portal opens soon.

For testing, USE_TENDERLY_FORK is used to move the island mode to
`default` on load. We may want to introduce a separate flag for this.