Add workflow creating PR from `main` to `stage-live`
The idea is thet whenever some PR gets merged to `main` we want to auto-create
another PR that updates `stage-live` branch with changes from `main`. If there's
already a PR open between those two branches, we want it to be updated
with the latest `main` changes.
In next commits we'll add more jobs (syncing different deployment-related
branches) to the workflow.