Rewrite XP glossary lookup
The previous solution had an edge case where, if the target address was
in the last file in the glossary, we would instead look for them in the
first file and fail to find the claim.

We transition here from `findIndex` to `find`. We stop looking when the
*next* entry in the glossary has a startAddress > the target. We also
stop looking at the last entry in the glossary regardless, and always
check the last entry if no entry is found whose sibling has start >
target.

This *should* cover all corner cases correctly.