Use `thesis-valkyrie` user's acces token to create release PR
We had issues with triggering the `test-list.yml` workflow. It turned out this
was because GitHub prevents from triggering workflows by events triggered by the
`GITHUB_TOKEN` (and this was our case). To work around that we need to change
the token used for creating the release PRs to a different one that
`GITHUB_TOKEN` (we've decided to use a personal access token of the
`thesis-valkyrie` user). We also needed to set `persist-credentials` parameter
to `false` in `actions/checkout@v3` action in order to not save the auth token to
the git config (thanks to this we'll be able to create PR with the specified PAT
and trigger `test-list.yml` automatically after that event).