Disable auto connection feature when there is only MM available
Web3onboard remembers last wallet that was used. If it was MM but MM
becomes unavailable then web3onboard will try to auto connect to MM
and will fail. So to avoid problem with attempting auto connection
to the wallet that is not available on the wallet lists let's just disable
autoConnect feature for Metamask for now.